devops-db-fix-demo/service/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
prometheus-client
requests
